refactor: switch to src/ layout (flat modules + packages); update pyproject + langgraph.json; remove old sp_oncall package; delete stray src/__init__.py
