+++ src/executor.py
@@ -1,11 +1,11 @@
 import asyncio
 from langchain_core.messages import HumanMessage
 
-from sp_oncall.mcp_node import mcp_node
-from sp_oncall.util.llm import load_chat_model
-from sp_oncall.configuration import Configuration
+from mcp_node import mcp_node
+from util.llm import load_chat_model
+from configuration import Configuration
 from prompts.network_executor import NETWORK_EXECUTOR_PROMPT
-from sp_oncall.schemas import GraphState, StepExecutionResult
+from schemas import GraphState, StepExecutionResult
 
 
 def llm_network_executor(graph_state: GraphState) -> GraphState:
