+++ src/planner.py
@@ -1,10 +1,10 @@
 from langchain_core.messages import SystemMessage
 
 from prompts.planner import PLANNER_PROMPT
-from sp_oncall.util.llm import load_chat_model
-from sp_oncall.util.plans import load_plan_data
-from sp_oncall.configuration import Configuration
-from sp_oncall.schemas import GraphState, PlannerOutput
+from util.llm import load_chat_model
+from util.plans import load_plan_data
+from configuration import Configuration
+from schemas import GraphState, PlannerOutput
 
 
 def planner_node(state: GraphState) -> GraphState:
