+++ src/reporter.py
@@ -2,9 +2,9 @@ import json
 from typing import Dict, Any
 from langchain_core.messages import SystemMessage
 
-from sp_oncall.schemas import GraphState
-from sp_oncall.util.llm import load_chat_model
-from sp_oncall.configuration import Configuration
+from schemas import GraphState
+from util.llm import load_chat_model
+from configuration import Configuration
 from prompts.report_generator import REPORT_GENERATOR_PROMPT_TEMPLATE
