+++ src/schemas/assessment_schema.py
